read.md
Lexical analysis
 the compiler is a program that accepts the source program as the input and produces the object
program. Since compiler is a very program, it is very difficult to understand the compiler design. So, the
designers of the compiler have divided the compilation of activities into various phases.
 Where lexical analysis is a first phase of a compiler is additionally called as scanner. Lexical breaks the
input into smallest significant sequence called tokens which parser uses for next subsequent section of a
compiler. Each token is a combination of characters that indicates a unit of information in the source code.
Examples of tokens such as identifiers, keywords, operators or special symbols, punctuation marks, etc